add Maia SDR to ant and e200
This adds Maia SDR to the ant and e200 patchsets. All the existing
ADI infrastructure is preserved, with maia-hdl piggybacking on the
same ADC inputs. This makes it possible to use Maia SDR and IIO at
the same time, with the only drawback of Maia SDR taking up 400 MiB
of reserved memory for the IQ recorder.

Signed-off-by: Daniel Estévez <daniel@destevez.net>